+++ src/main/java/com/hackathonteam7/mainprojectbackend/course/dto/CourseStatsResponse.java
@@ -1,7 +1,7 @@
 package com.hackathonteam7.mainprojectbackend.course.dto;
 
 public record CourseStatsResponse(
-        Long viewCount,
+        Integer viewCount,
         long participants,
         long completed,
         long abandoned,
